requirements.txt
amqp==5.2.0
ApiTestEngine==1.0.3
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-filter==24.2
django-simpleui==2024.4.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==13.4.0
idna==3.7
jsonpath==0.82
kombu==5.3.7
prompt_toolkit==3.0.46
pyasn1==0.6.0
PyJWT==2.8.0
PyMySQL==1.0.2
python-crontab==3.1.0
python-dateutil==2.9.0.post0
redis==5.0.4
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
